// src/features/movies/types/movie.types.ts
export interface TVShow {
  id: number
  name: string
  overview: string
  poster_path: string | null
  backdrop_path: string | null
  first_air_date: string
  vote_average: number
  vote_count: number
  genre_ids: number[]
  adult: boolean
  original_language: string
  original_name: string
  popularity: number
  origin_country: string[]
}
// ...
export interface TVShowDetails extends TVShow {
  genres: Genre[]
  episode_run_time: number[]
  status: string
  tagline: string
  homepage: string
  production_companies: ProductionCompany[]
  production_countries: ProductionCountry[]
  spoken_languages: SpokenLanguage[]
  number_of_episodes: number
  number_of_seasons: number
  seasons: Season[]
  networks: Network[]
  created_by: CreatedBy[]
  in_production: boolean
  languages: string[]
  last_air_date: string | null
  next_episode_to_air: Episode | null
  last_episode_to_air: Episode | null
  type: string
}

export interface Season {
  id: number
  air_date: string | null
  episode_count: number
  name: string
  overview: string
  poster_path: string | null
  season_number: number
  vote_average: number
}

export interface SeasonDetails extends Season {
  episodes: Episode[]
}

export interface Episode {
  id: number
  air_date: string | null
  episode_number: number
  name: string
  overview: string
  runtime: number | null
  season_number: number
  still_path: string | null
  vote_average: number
  vote_count: number
  production_code: string
  show_id: number
  crew: Crew[]
  guest_stars: Cast[]
}

export interface Network {
  id: number
  logo_path: string | null
  name: string
  origin_country: string
}

export interface NetworkDetails {
  id: number
  logo_path: string | null
  name: string
  origin_country: string
  headquarters: string
  homepage: string | null
}

export interface CreatedBy {
  id: number
  credit_id: string
  name: string
  gender: number
  profile_path: string | null
}

export interface Genre {
  id: number
  name: string
}

export interface ProductionCompany {
  id: number
  logo_path: string | null
  name: string
  origin_country: string
}

export interface ProductionCompanyDetails extends ProductionCompany {
  description: string
  headquarters: string
  homepage: string
  parent_company: ProductionCompany | null
}

export interface ProductionCountry {
  iso_3166_1: string
  name: string
}

export interface SpokenLanguage {
  iso_639_1: string
  name: string
  english_name: string
}
// ...
export interface Cast {
  id: number
  name: string
  character: string
  profile_path: string | null
  order: number
  cast_id: number
  credit_id: string
  adult: boolean
  gender: number
  known_for_department: string
  original_name: string
  popularity: number
}

export interface Crew {
  id: number
  name: string
  job: string
  department: string
  profile_path: string | null
  credit_id: string
  adult: boolean
  gender: number
  known_for_department: string
  original_name: string
  popularity: number
}